[DEV-21575] - Restructure Code - Restructured code to match HashiCorp scafolding project for SDKv2. - Added sevice directroy to group data sources and resources by product. Added cmp sub directory which now contains all data sources and resources for CMP. This structure will allow us to have a single provider for all CloudBolt products.
