+++ internal/service/cmp/data_source_cb_blueprint.go
@@ -1,4 +1,4 @@
-package cloudbolt
+package cmp
 
 import (
 	"fmt"
@@ -7,7 +7,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func dataSourceCloudBoltBlueprint() *schema.Resource {
+func DataSourceCloudBoltBlueprint() *schema.Resource {
 	return &schema.Resource{
 		Read: dataSourceCloudBoltBlueprintRead,
 
@@ -32,7 +32,7 @@ func dataSourceCloudBoltBlueprint() *schema.Resource {
 }
 
 func dataSourceCloudBoltBlueprintRead(d *schema.ResourceData, m interface{}) error {
-	apiClient := m.(Config).APIClient
+	apiClient := m.(*cbclient.CloudBoltClient)
 	name := d.Get("name").(string)
 	id := d.Get("id").(string)
