+++ main.go
@@ -4,13 +4,13 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	"github.com/cloudboltsoftware/terraform-provider-cloudbolt/cloudbolt"
+	"github.com/cloudboltsoftware/terraform-provider-cloudbolt/internal/provider"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
-			return cloudbolt.Provider()
+			return provider.Provider()
 		},
 	})
 }
